linker: remove unecessary usize casts

+++ src/image.rs
@@ -84,7 +84,7 @@ impl<'process> SharedObject<'process> {
         let dynamic = dyn::get_dynamic_array(load_bias as u64, phdrs).unwrap();
         let link_info = dyn::LinkInfo::new(&dynamic, load_bias);
         let num_syms = (link_info.strtab - link_info.symtab) / sym::SIZEOF_SYM;
-        let symtab = sym::from_raw(link_info.symtab as *const sym::Sym, num_syms as usize);
+        let symtab = sym::from_raw(link_info.symtab as *const sym::Sym, num_syms);
         let strtab = Strtab::from_raw(link_info.strtab as *const u8, link_info.strsz as usize);
         let libs = dyn::get_needed(dynamic, &strtab, link_info.needed_count);
         let relatab = rela::get(link_info.rela, link_info.relasz as usize, link_info.relaent as usize, link_info.relacount as usize);
@@ -128,7 +128,7 @@ impl<'process> SharedObject<'process> {
 
                 let link_info = dyn::LinkInfo::new(dynamic, load_bias as usize);
                 // TODO: swap out the link_info syment with compile time constant SIZEOF_SYM?
-                let num_syms = ((link_info.strtab - link_info.symtab) / link_info.syment) as usize; // this _CAN'T_ generally be valid; but rdr has been doing it and scans every linux shared object binary without issue... so it must be right!
+                let num_syms = ((link_info.strtab - link_info.symtab) / link_info.syment); // this _CAN'T_ generally be valid; but rdr has been doing it and scans every linux shared object binary without issue... so it must be right!
                 let symtab = sym::from_raw(link_info.symtab as *const sym::Sym, num_syms);
                 let strtab = Strtab::from_raw(link_info.strtab as *const u8, link_info.strsz);
                 let libs = dyn::get_needed(dynamic, &strtab, link_info.needed_count);
